+++ Assets/com.kaizerwald.formation/CHANGELOG.md
@@ -7,3 +7,9 @@ Add Function
 
 Added:
 UnorderedFormationMatrix<T>
+
+1.0.2
+
+Added:
+Ability to set Formation from FormationData without creating a new object
+Added to FormationMatrix and UnorderedFormationMatrix setter to CurrentFormation and TargetFormation from FormationData